requirements.txt
alembic==0.9.6
coverage==4.4.2
Mako==1.0.7
MarkupSafe==1.0
progressbar2==3.34.3
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-editor==1.0.3
python-utils==2.2.0
six==1.11.0
SQLAlchemy==1.1.15
